requirements.txt
hydra-core==1.2.0
imageio==2.22.0
opencv-python==4.6.0.66
pydantic==1.9.1
pytorch-lightning==1.6.3
setuptools==58.2.0
torchvision==0.11.3
